conversion

+++ single_include/nadjieb/mjpeg_streamer.hpp
@@ -311,7 +311,7 @@ static int readFromSocket(SocketFD socket, char* buffer, size_t length, int flag
 #ifdef NADJIEB_MJPEG_STREAMER_PLATFORM_WINDOWS
     return ::recv(socket, buffer, (int)length, flags);
 #else
-    return ::recv(socket, buffer, (int)length, flags);
+    return ::recv((int)socket, buffer, (int)length, flags);
 #endif
 }
 
@@ -319,7 +319,7 @@ static int sendViaSocket(SocketFD socket, const char* buffer, size_t length, int
 #ifdef NADJIEB_MJPEG_STREAMER_PLATFORM_WINDOWS
     return ::send(socket, buffer, (int)length, flags);
 #else
-    return ::send(socket, buffer, (int)length, flags);
+    return ::send((int)socket, buffer, (int)length, flags);
 #endif
 }
